browser-use-dev/requirements.txt
fastapi~=0.116.0
uvicorn[standard]
python-dotenv~=1.1.1
browser-use==0.5.11
openai==1.99.2
pillow
aiohttp>=3.9.0
websockets
docker
requests==2.32.3
langchain-openai==0.3.27
pydantic==2.11.7
pydantic-settings>=2.5.2
motor==3.1.1
pymongo==4.3.3
PyJWT
starlette~=0.46.2
capsolver
psutil~=7.0.0
itsdangerous>=2.1.2